% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.jump_test}
\alias{print.jump_test}
\title{Print a jump-test result}
\usage{
\method{print}{jump_test}(x, ...)
}
\arguments{
\item{x}{A \code{jump_test} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a jump-test result
}
